// SourceCode/Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/shared/models/category.model.ts
export class CategoryModel {
  CategoryId: string;
  CategoryName: string;
  CategoryCode: string;
  CategoryTypeId: string;
  CreatedById: string;
  CreatedDate: Date;
  UpdatedById: string;
  UpdatedDate: string;
  Active: Boolean;
  IsEdit: Boolean;
  IsDefault: Boolean;
  StatusName: string;
  PotentialName: string;
  CategoryTypeName: string;
  constructor() { }
}